# Bressolium/backend/BressoliumProject/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bressolium API",
        "description": "API REST del videojuego de estrategia colaborativa Bressolium. Todos los endpoints devuelven JSON con el formato unificado {success, data, error}.",
        "contact": {
            "name": "Equipo Bressolium"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/game/{gameId}/close-round": {
            "post": {
                "tags": [
                    "Round"
                ],
                "summary": "Cierra la jornada actual y genera la siguiente",
                "description": "Resuelve los votos de tecnologías e inventos, activa los más votados, produce materiales según las casillas exploradas, marca jugadores AFK y crea la siguiente jornada.",
                "operationId": "fce22117f788fe97b2e86669d116ac38",
                "parameters": [
                    {
                        "name": "gameId",
                        "in": "path",
                        "description": "UUID de la partida",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jornada cerrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Jornada cerrada correctamente."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No perteneces a esta partida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Partida no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/tiles/{id}/explore": {
            "post": {
                "tags": [
                    "Tile"
                ],
                "summary": "Explora una casilla",
                "description": "Marca una casilla como explorada por el usuario. Consume una acción de la jornada actual. La casilla debe ser adyacente a otra ya explorada por el jugador.",
                "operationId": "ddf964b7a5daaf7772206e4ef46fafb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la casilla a explorar",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Casilla explorada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "coord_x": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "coord_y": {
                                                    "type": "integer",
                                                    "example": 7
                                                },
                                                "explored": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "type": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "El usuario no pertenece a esta partida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Casilla ya explorada / no adyacente / sin acciones / casilla pueblo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/tiles/{id}/upgrade": {
            "post": {
                "tags": [
                    "Tile"
                ],
                "summary": "Mejora una casilla al siguiente nivel",
                "description": "Sube de nivel el tipo de la casilla. Requiere que la casilla esté explorada, que existan los materiales necesarios en la partida y, en algunos casos, que la tecnología requerida esté activa. Consume una acción.",
                "operationId": "e9e4572af6ffb3c46b1444812f508cf3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la casilla a mejorar",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Casilla mejorada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "El usuario no pertenece a esta partida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Materiales insuficientes / tecnología no disponible / sin acciones / casilla no explorada / sin niveles superiores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "description": "Formato unificado de respuesta de la API",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    },
                    "error": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ApiError": {
                "description": "Respuesta de error estándar",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "error": {
                        "type": "string",
                        "example": "Mensaje de error"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de Laravel Sanctum obtenido al hacer login. Se envía en la cabecera Authorization: Bearer {token}.",
                "bearerFormat": "Sanctum-Token",
                "scheme": "bearer"
            }
        }
    },
...
}